refactoring web helpers and tests
