+++ app/app.rb
@@ -1,7 +1,7 @@
 ENV['RACK_ENV'] ||= 'development'
 
 require 'sinatra/base'
-require './app/models/database_setup.rb'
+require './database_setup.rb'
 
 class DwellBNB < Sinatra::Base
