+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ ENV['RACK_ENV'] = 'test'
 
 require './app/app'
 require './app/models/user'
-require './spec/helpers/user_helper'
+require './spec/helpers/web_helpers'
 require 'capybara'
 require 'capybara/rspec'
 require 'database_cleaner'
@@ -21,7 +21,6 @@ SimpleCov.start
 
 RSpec.configure do |config|
 
-  config.include UserHelpers
 
   config.before(:suite) do
     DatabaseCleaner.strategy = :transaction
